Move validator to lint
