+++ test/mercury/resource-validation.test.ts
@@ -6,11 +6,10 @@
  */
 /* eslint-disable no-undef */
 "use strict";
-import { validateFile } from "../../src/validator";
+import { validateFile } from "../../src/lint/lint";
 import {
   getHappySpec,
   conforms,
-  breaksOnlyOneRule,
   breaksTheseRules,
   renameKey,
   renderSpecAsFile
